+++ tool/context.go
@@ -15,26 +15,16 @@
 package tool
 
 import (
-	"github.com/google/uuid"
-
 	"google.golang.org/adk/agent"
 	"google.golang.org/adk/session"
 	"google.golang.org/adk/tool/toolconfirmation"
 )
 
-// NewToolContext constructs a Context for a tool execution.
+// NewToolContext constructs a ToolContext for a tool execution.
 //
-// If functionCallID is empty a new UUID is generated. If actions is nil a
-// fresh session.EventActions with empty StateDelta and ArtifactDelta is
-// allocated. The returned Context is backed by the same *callbackContext
-// implementation used for CallbackContext, so all callback-context semantics
-// (state delta tracking, artifact delta tracking, etc.) apply.
+// Deprecated: use agent.NewToolContext directly. This wrapper exists only
+// to minimize churn during the migration and will be removed in a future
+// release.
 func NewToolContext(ic agent.InvocationContext, functionCallID string, actions *session.EventActions, confirmation *toolconfirmation.ToolConfirmation) ToolContext {
-	if functionCallID == "" {
-		functionCallID = uuid.NewString()
-	}
-	// agent.NewToolCallbackContext returns *agent.callbackContext (an
-	// unexported type). Assigning it to the Context interface here is the
-	// compile-time assertion that the method set satisfies tool.Context.
-	return agent.NewToolCallbackContext(ic, functionCallID, actions, confirmation)
+	return agent.NewToolContext(ic, functionCallID, actions, confirmation)
 }
